refactor(common): relocate config/exceptions/logging/_version under common/ (#164)
Phase 8e of the src/srunx/ cleanup (#156). Moves the foundation-layer
modules into a dedicated ``src/srunx/common/`` package so the
architecture guard's ``common`` layer has a physical home matching its
conceptual one.

Moves:
  - src/srunx/config.py         → src/srunx/common/config.py
  - src/srunx/exceptions.py     → src/srunx/common/exceptions.py
  - src/srunx/logging.py        → src/srunx/common/logging.py
  - src/srunx/_version.py       → src/srunx/common/_version.py
    (auto-generated by hatch-vcs; pyproject.toml ``version-file``
    updated; .gitignore updated to match)

Shim strategy:
  - Each top-level shim file (``srunx/config.py`` etc.) does
    ``from srunx.common.X import *`` so static type-checkers resolve
    names correctly, THEN aliases ``sys.modules[__name__] = canonical``
    at runtime so the shim and canonical refer to the SAME module
    object. This preserves monkey-patching (``srunx.config._config =
    None``), ``isinstance`` checks, and attribute identity across the
    two paths.

Internal update:
  - ``srunx.sync.lock`` now imports ``_user_config_dir`` directly from
    ``srunx.common.config`` (private names aren't picked up by
    ``from X import *``).

Architecture guard:
  - ``common`` added to ``MODULE_LAYERS`` alongside the top-level
    shim names. Existing cross-layer rules unchanged.

Behavior unchanged. 1977 of 1981 tests pass; the 4 failing tests are
env-specific pre-existing issues (local ``~/.config/srunx/config.json``
overrides defaults) that fail on main too.
