+++ .gitignore
@@ -465,6 +465,9 @@ $RECYCLE.BIN/
 .coverage
 coverage.xml
 # Auto-generated by hatch-vcs
+src/srunx/common/_version.py
+# Legacy location (relocated in #164 Phase 8e); ignore in case an older
+# build left a stray file behind.
 src/srunx/_version.py
 
 # Playwright
