+++ pyproject.toml
@@ -66,7 +66,7 @@ include = [
 source = "vcs"
 
 [tool.hatch.build.hooks.vcs]
-version-file = "src/srunx/_version.py"
+version-file = "src/srunx/common/_version.py"
 
 [dependency-groups]
 dev = [
