+++ src/srunx/__init__.py
@@ -1,7 +1,7 @@
 """srunx - Python library for SLURM job management."""
 
 try:
-    from srunx._version import __version__
+    from srunx.common._version import __version__
 except ImportError:
     __version__ = "0.0.0+unknown"
 __author__ = "ksterx"
